requirements.txt
flask
flask-socketio
werkzeug
rich
PyYAML
